+++ src/components/PlayArea/spec.tsx
@@ -1,8 +1,10 @@
 import { mount } from 'enzyme';
 import React from 'react';
+import { act } from 'react-dom/test-utils';
 import renderer from 'react-test-renderer';
 import 'jest-styled-components';
 import Card from '@components/Card';
+import Deck from '@components/Deck';
 import PlayArea from '@components/PlayArea';
 
 describe('PlayArea', () => {
@@ -20,6 +22,27 @@ describe('PlayArea', () => {
     it('should render the correct amount of Card components', () => {
       const wrapper = mount(PlayAreaMock);
       expect(wrapper.find(Card)).toHaveLength(10);
+    });
+  });
+
+  describe('Deck', () => {
+    it('should render the Deck component', () => {
+      const wrapper = mount(PlayAreaMock);
+      expect(wrapper.find(Deck)).toHaveLength(1);
+    });
+
+    it('should should initially contain the "Draw a Card." paragraph', () => {
+      const wrapper = mount(PlayAreaMock);
+      expect(wrapper.find(Deck).find('p')).toHaveLength(1);
+    });
+
+    it('should not contain a "Draw a card." paragraph tag after being clicked', () => {
+      const wrapper = mount(PlayAreaMock);
+      const selector = wrapper.find('.deck').at(0).prop('onClick');
+      const event = {} as React.MouseEvent;
+      if (selector) act(() => selector(event));
+      wrapper.update();
+      expect(wrapper.find('p')).toHaveLength(0);
     });
   });
 });