+++ docs/changelog.md
@@ -13,6 +13,7 @@ Ends August 17, 2020
 - rename `handleCardClick` to `handleCardMouseDown`.
 - memoize Deck component.
 - pass in title prop to Card components.
+- add "Draw a Card." experience to Deck component.
 
 ## Sprint 1
 Ended August 3, 2020